backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.0,<5.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Environment
python-dotenv>=1.0,<2.0
python-decouple>=3.8,<4.0

# PDF/Document Parsing
PyMuPDF>=1.23,<2.0
python-docx>=1.1,<2.0
pytesseract>=0.3,<1.0

# NLP & Resume Parsing
spacy>=3.7,<4.0
nltk>=3.8,<4.0

# ML & Deep Learning
tensorflow>=2.15,<3.0
opencv-python>=4.9,<5.0
mediapipe>=0.10,<1.0
numpy>=1.26
Pillow>=10.2,<11.0

# WebSocket & Async
websockets>=12.0,<13.0
aiohttp>=3.9,<4.0
httpx>=0.26,<1.0

# Google AI
google-generativeai>=0.4,<1.0

# Utilities
python-dateutil>=2.8,<3.0
pydantic>=2.5,<3.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
pytest-asyncio>=0.23,<1.0

# Production
gunicorn>=21.0,<22.0
uvicorn>=0.27,<1.0
whitenoise>=6.6,<7.0
